requirements.txt
torch
torchvision
progressbar2==4.2.0
tqdm